Add better equality checks (can now solve more puzzles)
Since the order of the tubes doesn't matter, two states can be
considered equal if all the contents of each of their tubes are the
same. This also prunes many additional branches in the BFS, since it was
searching through many trees that essentially represented the same
state. With this change, the solver was able to solve the 3 puzzles that
it was stuck on before.